requirements.txt
asgiref==3.8.1
distlib==0.3.6
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
filelock==3.12.0
packaging==23.2
platformdirs==3.5.1
psycopg2==2.9.7
pytz==2023.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.23.0







